Validação automática de schemas de modelos (#73)
* Adiciona função para inferir o schema do BigQuery de um modelo amora

* Adiciona documentação

* Difere `Column` de `ColumnElement` em `amora.models`

* Corrige schemas e adiciona documentação a modelos de exemplo/teste

* Adiciona testes genéricos, que podem ser reutilizados por múltiplos projetos

* Aumenta cobertura de testes

+++ examples/amora_project/models/heart_agg.py
@@ -11,7 +11,7 @@ class HeartRateAgg(AmoraModel, table=True):
 
     avg: float
     sum: float
-    count: float
+    count: int
     year: int = Field(primary_key=True)
     month: int = Field(primary_key=True)
